added children example
